% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amelie.R
\docType{package}
\name{amelie}
\alias{amelie}
\alias{amelie-package}
\title{amelie: A package for anomaly detection.}
\description{
Anomaly detection with maximum likelihood estimates and normal probability functions.
}
\section{Amelie functions}{

The package contains a function for running the anomaly detection algorithm.
}

\section{More information}{

\code{\link{ad}} documents the main \code{ad} function.

For more details and examples, see the vignette.
}

