% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdfunc.R
\name{pdfunc}
\alias{pdfunc}
\title{pdfunc: compute probability density function.}
\usage{
pdfunc()
}
\arguments{
\item{data}{A data frame containing the features (predictors).}

\item{x}{A matrix of numeric features.
assumed to be anomalous.}

\item{univariate}{Logical indicating whether the univariate pdf should be computed.}

\item{na.action}{A function specifying the action to be taken if NAs are
found.}

\item{...}{Currently not used.}
}
\value{
A list containing the following:
  \item{univariate}{TODO}
  \item{x_mean}{TODO}
  \item{x_sd}{TODO}
  \item{probs}{TODO}
}
\description{
pdfunc: compute probability density function.
}
\details{
\code{pdfunc} computes univariate or multivariate probabilities for a set of observations.
}
\references{
\url{https://www.coursera.org/learn/machine-learning}
}
