% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.anode.R
\name{predict.anode}
\alias{predict.anode}
\title{Predict Method for anode Objects}
\usage{
\method{predict}{anode}(object, newdata, na.action, ...)
}
\arguments{
\item{object}{An object of class \code{anode}, created by the function \code{anode}.}

\item{newdata}{A data frame or matrix containing new data.}

\item{na.action}{TODO}

\item{...}{Currently not used.}
}
\value{
A vector of predicted values.
}
\description{
Predict Method for anode Objects
}
\details{
Details go here.
}
\examples{
# Examples go here.

}
