% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anode.R
\name{anode}
\alias{anode}
\alias{anode.formula}
\alias{anode.default}
\alias{print.anode}
\title{Anomaly Detection}
\usage{
anode(x, ...)

\method{anode}{formula}(formula, data, na.action = na.omit)

\method{anode}{default}(x, y, na.action = na.omit)

\method{print}{anode}(x, ...)
}
\arguments{
\item{formula}{An object of class "formula": a symbolic description of the
model to be fitted.}

\item{data}{a data frame containing the features (predictors) and target.}
}
\value{
An object of class \code{anode}:
  \item{call}{The original call to \code{anode}.}
  \item{epsilon}{The threshold value.}
}
\description{
Anomaly Detection

#'@export
anode <-
  function (x, ...)
    UseMethod ("anode")
}
\details{
Details go here.
Uses F1 score.
Features are assumed to be continuous, and target is assumed to be either
0 or 1. NAs not supported.
}
\examples{
# Examples go here.

}
