% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anode.R
\name{anode}
\alias{anode}
\alias{anode.formula}
\alias{anode.default}
\alias{print.anode}
\title{anode: supervised anomaly detection with normal probability
functions.}
\usage{
anode(x, ...)

\method{anode}{formula}(formula, data, na.action = na.omit, ...)

\method{anode}{default}(x, y, na.action = na.omit, ...)

\method{print}{anode}(x, ...)
}
\arguments{
\item{x}{A matrix of numeric features.}

\item{...}{Currently not used.}

\item{formula}{An object of class "formula": a symbolic description of the
model to be fitted.}

\item{data}{A data frame containing the features (predictors) and target.}

\item{na.action}{A function specifying the action to be taken if NAs are
found.}

\item{y}{A matrix of numeric predicted values, either 0 or 1, with 1
assumed to be anomalous.}
}
\value{
An object of class \code{anode}:
  \item{call}{The original call to \code{anode}.}
  \item{epsilon}{The threshold value.}
}
\description{
anode: supervised anomaly detection with normal probability
functions.
}
\details{
Details go here.

\code{anode} implements anomaly detection as described in the Coursera Machine
Learning course.

The approach is set up as a binary classification problem. Features are
assumed to be continuous, and the target is assumed to take on values of 0
(negative case, no anomaly) or 1 (positive case, anomaly).

The threshold \code{epsilon} is optimized using the F1 score.

NAs are supported in the following way: TODO EXPLANATION.
}
\examples{
# Examples go here.

}
