% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdfunc.R
\name{pdfunc}
\alias{pdfunc}
\title{pdfunc: compute the probability density function.}
\usage{
pdfunc(x, univariate = TRUE)
}
\arguments{
\item{x}{A matrix of numeric features.
assumed to be anomalous.}

\item{univariate}{Logical indicating whether the univariate pdf should be computed.}
}
\value{
A list containing the following:
  \item{univariate}{Logical indicating which pdf was computed.}
  \item{x_mean}{Means of the features.}
  \item{x_sd}{Standard deviations of the features.}
  \item{probs}{values of the density function.}
}
\description{
pdfunc: compute the probability density function.
}
\details{
\code{pdfunc} computes univariate or multivariate probabilities for a set of observations.

Standard deviations are computed using \code{sd}, where denominator
\code{n-1} is used (sample standard deviation).
}
